pyproject.toml
[project]
name = "safa"
version = "0.1.0"
description = "A command line tool for running SAFA in the command line."
authors = [
    { name = "Alberto Rodriguez", email = "alberto@safa.ai.com" }
]
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }

dependencies = [
    "GitPython==3.1.43",
    "anthropic==0.30.1",
    "certifi==2024.6.2",
    "langchain-anthropic==0.1.19",
    "langchain-chroma==0.1.2",
    "langchain-community==0.2.6",
    "langchain-core==0.2.11",
    "langchain-huggingface==0.0.3",
    "langchain-text-splitters==0.2.2",
    "langchain==0.2.6",
    "python-dotenv==1.0.1",
    "requests==2.32.3",
    "sentence-transformers==3.0.1"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["safa"]

[project.scripts]
safa = "safa.runner:main"